test env graceful cleanup & slo suite cleanup
